requirements.txt
Flask>=3.1.0
flask-cors>=5.0.0
scikit-learn>=1.7.0
nltk>=3.9.1
pdfplumber>=0.11.4
python-docx>=1.1.2
sentence-transformers>=2.7.0
torch>=2.1.0
